Bug #1257: compile warnings due to pid_t not being an int
Forward port of 2.5 patch and added use of debugs() where applicable

+++ helpers/ntlm_auth/SMB/ntlm_auth.c
@@ -328,7 +328,7 @@ manage_request()
 	    alarm(0);
 	    signal(SIGALRM, SIG_DFL);
 	    if (got_timeout != 0) {
-		fprintf(stderr, "ntlm-auth[%d]: Timeout during authentication.\n", getpid());
+		fprintf(stderr, "ntlm-auth[%ld]: Timeout during authentication.\n", (long)getpid());
 		SEND("BH Timeout during authentication");
 		got_timeout = 0;
 		return;
